refactor: Rename package "trainining" to training
