+++ app/src/main/java/com/haidoan/android/stren/feat/training/exercises/ExercisesUiState.kt
@@ -1,0 +1,8 @@
+package com.haidoan.android.stren.feat.training.exercises
+
+import com.haidoan.android.stren.core.model.Exercise
+
+internal sealed interface ExercisesUiState {
+    object Loading : ExercisesUiState
+    data class LoadComplete(val exercises: List<Exercise>) : ExercisesUiState
+}
