--- app/src/main/java/com/haidoan/android/stren/feat/trainining/routines/add_edit/AddEditRoutineUiState.kt
@@ -1,9 +1,0 @@
-package com.haidoan.android.stren.feat.trainining.routines.add_edit
-
-import com.haidoan.android.stren.core.model.TrainedExercise
-
-internal sealed interface AddEditRoutineUiState {
-    object Loading : AddEditRoutineUiState
-    object EmptyRoutine : AddEditRoutineUiState
-    data class IsEditing(val trainedExercises: List<TrainedExercise>) : AddEditRoutineUiState
-}
