+++ app/src/test/java/com/haidoan/android/stren/feat/training/exercises/ExerciseDetailViewModelTest.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.SavedStateHandle
 import com.haidoan.android.stren.core.repository.fake.FakeExercisesRepository
 import com.haidoan.android.stren.core.testing.data.EXERCISES_TEST_DATA
 import com.haidoan.android.stren.core.testing.util.MainDispatcherRule
-import com.haidoan.android.stren.feat.trainining.exercises.detail.EXERCISE_ID_ARG
-import com.haidoan.android.stren.feat.trainining.exercises.detail.ExerciseDetailViewModel
+import com.haidoan.android.stren.feat.training.exercises.detail.EXERCISE_ID_ARG
+import com.haidoan.android.stren.feat.training.exercises.detail.ExerciseDetailViewModel
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.collect
